+++ lib/DisplayPatternLib/src/DisplayPatterns/FireDisplayPattern.h
@@ -31,7 +31,7 @@ class FireDisplayPattern : public DisplayPattern
         FirePatternType m_patternType{FirePatternType::Solid};
         byte m_sparking{120};
         byte m_cooling{50};
-        ulong m_heatColors[256];
+        uint m_heatColors[256];
         
         // Maintains the current "heat" level for each row grouping.
         std::vector<std::vector<int>> m_combinedRowHeats;
